完成序列化代码审计，修正除后3项外的测试用例

+++ data/TablePurchaseRecord.h
@@ -20,7 +20,7 @@ typedef struct {
     int partId;
     int sellerId;
     int amount;
-    float total;
+    double total;
     int orderId;
     enum PurchaseRecordStatus status;
     Time time;
